deploy/deploy_requirements.txt
# currently installed in conda env named: fraud-deploy-env
google-cloud-pubsub
google-cloud-bigquery
google-cloud-aiplatform
pandas==2.3.1
numpy==2.3.2
scikit-learn==1.7.1
tqdm==4.67.1
imbalanced-learn==0.14.0
pyarrow==21.0.0